# async_api/OpenApi.yaml
openapi: 3.0.0
info:
  title: Movies ListApi
  version: v1
servers:
  - url: http://127.0.0.1:9000
paths:
...
  /api/v1/genres/:
    get:
      description: Возвращает список жанров
      parameters:
        - name: page[size]
          in: query
          description: Количество элеменнтов на странице
          required: false
          schema:
            type: integer
            default: 50
        - name: page[number]
          in: query
          description: Номер страницы
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: 'Возвращает список фильмов с полями `id`, `title`, `imdb_rating`'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Genre"
        '422':
          description: "Validation Error"
          content: 
            application/json: 
              schema: 
                $ref: "#/components/schemas/ValidationError"
      summary: "Get Genres"
      tags: 
        - Genres
  /api/v1/genres/{genre_id}/:
    get:
      description: "Возвращает информацию о жанре"
      parameters:
        - in: path
          name: genre_id  
          required: true
          schema:
            type: string
            format: uuid
          description: ID жанра
      responses:
        "200":
          description: "Возвращает информацию о фильме"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Genre"
        '422':
          description: "Validation Error"
          content: 
            application/json: 
              schema: 
                $ref: "#/components/schemas/ValidationError"
      summary: "Genre Details"
      tags: 
        - Genres
...
components:
  schemas:
...
    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID
        name:
          type: string
          description: Название жанра
          example: Comedy
    ValidationError: 
      properties: 
        loc: 
          items: 
            anyOf: 
              - 
                type: string
              - 
                type: integer
          title: Location
          type: array
        msg: 
          title: Message
          type: string
        type: 
          title: "Error Type"
          type: string
      required: 
        - loc
        - msg
        - type
      title: ValidationError
      type: object
...